`ColumnSizeControl` move to column components
