+++ src/blocks/column-maxi/components/index.js
@@ -1,0 +1,2 @@
+// eslint-disable-next-line import/prefer-default-export
+export { default as ColumnSizeControl } from './column-size-control';
