+++ src/blocks/column-maxi/inspector.js
@@ -7,11 +7,8 @@ import { InspectorControls } from '@wordpress/block-editor';
 /**
  * Internal dependencies
  */
-import {
-	AccordionControl,
-	ColumnSizeControl,
-	SettingTabsControl,
-} from '../../components';
+import { AccordionControl, SettingTabsControl } from '../../components';
+import { ColumnSizeControl } from './components';
 import * as inspectorTabs from '../../components/inspector-tabs';
 import { customCss } from './data';
 import ResponsiveTabsControl from '../../components/responsive-tabs-control';
